[Task Manager] allow multiple task claiming strategiess
see https://github.com/elastic/kibana/issues/155770

Make the task manager task claiming algorithm selectable, to allow
alternative implementations in the future.  No other implementations
are provided here, this is setup for adding the next one.
