+++ x-pack/plugins/task_manager/server/config.test.ts
@@ -13,6 +13,7 @@ describe('config validation', () => {
     expect(configSchema.validate(config)).toMatchInlineSnapshot(`
       Object {
         "allow_reading_invalid_state": true,
+        "claim_strategy": "default",
         "ephemeral_tasks": Object {
           "enabled": false,
           "request_capacity": 10,
@@ -72,6 +73,7 @@ describe('config validation', () => {
     expect(configSchema.validate(config)).toMatchInlineSnapshot(`
       Object {
         "allow_reading_invalid_state": true,
+        "claim_strategy": "default",
         "ephemeral_tasks": Object {
           "enabled": false,
           "request_capacity": 10,
@@ -129,6 +131,7 @@ describe('config validation', () => {
     expect(configSchema.validate(config)).toMatchInlineSnapshot(`
       Object {
         "allow_reading_invalid_state": true,
+        "claim_strategy": "default",
         "ephemeral_tasks": Object {
           "enabled": false,
           "request_capacity": 10,
@@ -243,5 +246,14 @@ describe('config validation', () => {
     expect(() => {
       configSchema.validate(config);
     }).not.toThrowError();
+  });
+
+  test('the claim strategy is validated', () => {
+    const config = { claim_strategy: 'invalid-strategy' };
+    expect(() => {
+      configSchema.validate(config);
+    }).toThrowErrorMatchingInlineSnapshot(
+      `"The claim strategy is invalid: Unknown task claiming strategy (invalid-strategy)"`
+    );
   });
 });
