+++ x-pack/plugins/task_manager/server/config.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema, TypeOf } from '@kbn/config-schema';
+import { getTaskClaimer } from './task_claimers';
 
 export const MAX_WORKERS_LIMIT = 100;
 export const DEFAULT_MAX_WORKERS = 10;
@@ -24,6 +25,8 @@ export const DEFAULT_METRICS_RESET_INTERVAL = 30 * 1000; // 30 seconds
 
 // At the default poll interval of 3sec, this averages over the last 15sec.
 export const DEFAULT_WORKER_UTILIZATION_RUNNING_AVERAGE_WINDOW = 5;
+
+export const CLAIM_STRATEGY_DEFAULT = 'default';
 
 export const taskExecutionFailureThresholdSchema = schema.object(
   {
@@ -152,6 +155,7 @@ export const configSchema = schema.object(
       max: 100,
       min: 1,
     }),
+    claim_strategy: schema.string({ defaultValue: CLAIM_STRATEGY_DEFAULT }),
   },
   {
     validate: (config) => {
@@ -161,6 +165,11 @@ export const configSchema = schema.object(
         config.monitored_stats_required_freshness < config.poll_interval
       ) {
         return `The specified monitored_stats_required_freshness (${config.monitored_stats_required_freshness}) is invalid, as it is below the poll_interval (${config.poll_interval})`;
+      }
+      try {
+        getTaskClaimer(config.claim_strategy);
+      } catch (err) {
+        return `The claim strategy is invalid: ${err.message}`;
       }
     },
   }
