+++ x-pack/plugins/task_manager/server/ephemeral_task_lifecycle.test.ts
@@ -85,6 +85,7 @@ describe('EphemeralTaskLifecycle', () => {
           max_attempts: 20,
         },
         metrics_reset_interval: 3000,
+        claim_strategy: 'default',
         ...config,
       },
       elasticsearchAndSOAvailability$,
