+++ x-pack/plugins/task_manager/server/integration_tests/managed_configuration.test.ts
@@ -80,6 +80,7 @@ describe('managed configuration', () => {
         max_attempts: 20,
       },
       metrics_reset_interval: 3000,
+      claim_strategy: 'default',
     });
     logger = context.logger.get('taskManager');
