+++ x-pack/plugins/task_manager/server/lib/calculate_health_status.test.ts
@@ -57,6 +57,7 @@ const config = {
     max_attempts: 20,
   },
   metrics_reset_interval: 3000,
+  claim_strategy: 'default',
 };
 
 const getStatsWithTimestamp = ({
