+++ x-pack/plugins/task_manager/server/metrics/create_aggregator.test.ts
@@ -73,6 +73,7 @@ const config: TaskManagerConfig = {
   },
   version_conflict_threshold: 80,
   worker_utilization_running_average_window: 5,
+  claim_strategy: 'default',
 };
 
 describe('createAggregator', () => {
