+++ x-pack/plugins/task_manager/server/monitoring/configuration_statistics.test.ts
@@ -53,6 +53,7 @@ describe('Configuration Statistics Aggregator', () => {
         max_attempts: 20,
       },
       metrics_reset_interval: 3000,
+      claim_strategy: 'default',
     };
 
     const managedConfig = {
