+++ x-pack/plugins/task_manager/server/plugin.test.ts
@@ -78,6 +78,7 @@ const pluginInitializerContextParams = {
     max_attempts: 20,
   },
   metrics_reset_interval: 3000,
+  claim_strategy: 'default',
 };
 
 describe('TaskManagerPlugin', () => {
