+++ x-pack/plugins/task_manager/server/queries/task_claiming.ts
@@ -8,41 +8,30 @@
 /*
  * This module contains helpers for managing the task manager storage layer.
  */
-import apm from 'elastic-apm-node';
-import minimatch from 'minimatch';
-import { Subject, Observable, from, of } from 'rxjs';
-import { map, mergeScan } from 'rxjs/operators';
-import { groupBy, pick, isPlainObject } from 'lodash';
+import { Subject, Observable, of } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { groupBy, isPlainObject } from 'lodash';
 
 import { Logger } from '@kbn/core/server';
 
 import { asOk, asErr, Result } from '../lib/result_type';
 import { ConcreteTaskInstance } from '../task';
-import { TaskClaim, asTaskClaimEvent, startTaskTimer, TaskTiming } from '../task_events';
-import { shouldBeOneOf, mustBeAllOf, filterDownBy, matchesClauses } from './query_clauses';
+import { TaskClaim } from '../task_events';
 
-import {
-  updateFieldsAndMarkAsFailed,
-  IdleTaskWithExpiredRunAt,
-  InactiveTasks,
-  RunningOrClaimingTaskWithExpiredRetryAt,
-  SortByRunAtAndRetryAt,
-  tasksClaimedByOwner,
-  tasksOfType,
-  EnabledTask,
-} from './mark_available_tasks_as_claimed';
 import { TaskTypeDictionary } from '../task_type_dictionary';
-import {
-  correctVersionConflictsForContinuation,
-  TaskStore,
-  UpdateByQueryResult,
-  SearchOpts,
-} from '../task_store';
+import { TaskStore, UpdateByQueryResult } from '../task_store';
 import { FillPoolResult } from '../lib/fill_pool';
-import { TASK_MANAGER_TRANSACTION_TYPE } from '../task_running';
+import {
+  TaskClaimerOpts,
+  TaskClaimerFn,
+  ClaimOwnershipResult,
+  getTaskClaimer,
+} from '../task_claimers';
 
+export type { ClaimOwnershipResult } from '../task_claimers';
 export interface TaskClaimingOpts {
   logger: Logger;
+  strategy: string;
   definitions: TaskTypeDictionary;
   unusedTypes: string[];
   taskStore: TaskStore;
@@ -67,31 +56,25 @@ export interface FetchResult {
   docs: ConcreteTaskInstance[];
 }
 
-export interface ClaimOwnershipResult {
-  stats: {
-    tasksUpdated: number;
-    tasksConflicted: number;
-    tasksClaimed: number;
-  };
-  docs: ConcreteTaskInstance[];
-  timing?: TaskTiming;
+export function isClaimOwnershipResult(result: unknown): result is ClaimOwnershipResult {
+  return (
+    isPlainObject((result as ClaimOwnershipResult).stats) &&
+    Array.isArray((result as ClaimOwnershipResult).docs)
+  );
 }
-export const isClaimOwnershipResult = (result: unknown): result is ClaimOwnershipResult =>
-  isPlainObject((result as ClaimOwnershipResult).stats) &&
-  Array.isArray((result as ClaimOwnershipResult).docs);
 
-enum BatchConcurrency {
+export enum BatchConcurrency {
   Unlimited,
   Limited,
 }
 
-type TaskClaimingBatches = Array<UnlimitedBatch | LimitedBatch>;
-interface TaskClaimingBatch<Concurrency extends BatchConcurrency, TaskType> {
+export type TaskClaimingBatches = Array<UnlimitedBatch | LimitedBatch>;
+export interface TaskClaimingBatch<Concurrency extends BatchConcurrency, TaskType> {
   concurrency: Concurrency;
   tasksTypes: TaskType;
 }
-type UnlimitedBatch = TaskClaimingBatch<BatchConcurrency.Unlimited, Set<string>>;
-type LimitedBatch = TaskClaimingBatch<BatchConcurrency.Limited, string>;
+export type UnlimitedBatch = TaskClaimingBatch<BatchConcurrency.Unlimited, Set<string>>;
+export type LimitedBatch = TaskClaimingBatch<BatchConcurrency.Limited, string>;
 
 export const TASK_MANAGER_MARK_AS_CLAIMED = 'mark-available-tasks-as-claimed';
 
@@ -108,6 +91,7 @@ export class TaskClaiming {
   private readonly taskMaxAttempts: Record<string, number>;
   private readonly excludedTaskTypes: string[];
   private readonly unusedTypes: string[];
+  private readonly taskClaimer: TaskClaimerFn;
 
   /**
    * Constructs a new TaskStore.
@@ -120,12 +104,12 @@ export class TaskClaiming {
     this.maxAttempts = opts.maxAttempts;
     this.taskStore = opts.taskStore;
     this.getCapacity = opts.getCapacity;
-    this.logger = opts.logger;
+    this.logger = opts.logger.get('taskClaiming');
     this.taskClaimingBatchesByType = this.partitionIntoClaimingBatches(this.definitions);
     this.taskMaxAttempts = Object.fromEntries(this.normalizeMaxAttempts(this.definitions));
     this.excludedTaskTypes = opts.excludedTaskTypes;
     this.unusedTypes = opts.unusedTypes;
-
+    this.taskClaimer = getTaskClaimer(opts.strategy);
     this.events$ = new Subject<TaskClaim>();
   }
 
@@ -177,224 +161,43 @@ export class TaskClaiming {
     return this.events$;
   }
 
-  private emitEvents = (events: TaskClaim[]) => {
-    events.forEach((event) => this.events$.next(event));
-  };
-
   public claimAvailableTasksIfCapacityIsAvailable(
     claimingOptions: Omit<OwnershipClaimingOpts, 'size' | 'taskTypes'>
   ): Observable<Result<ClaimOwnershipResult, FillPoolResult>> {
     if (this.getCapacity()) {
-      return this.claimAvailableTasks(claimingOptions).pipe(
-        map((claimResult) => asOk(claimResult))
-      );
+      const opts: TaskClaimerOpts = {
+        batches: this.getClaimingBatches(),
+        claimOwnershipUntil: claimingOptions.claimOwnershipUntil,
+        taskStore: this.taskStore,
+        events$: this.events$,
+        getCapacity: this.getCapacity,
+        unusedTypes: this.unusedTypes,
+        definitions: this.definitions,
+        taskMaxAttempts: this.taskMaxAttempts,
+        excludedTaskTypes: this.excludedTaskTypes,
+      };
+      return this.taskClaimer(opts).pipe(map((claimResult) => asOk(claimResult)));
     }
     this.logger.debug(
       `[Task Ownership]: Task Manager has skipped Claiming Ownership of available tasks at it has ran out Available Workers.`
     );
     return of(asErr(FillPoolResult.NoAvailableWorkers));
   }
-
-  public claimAvailableTasks({
-    claimOwnershipUntil,
-  }: Omit<OwnershipClaimingOpts, 'size' | 'taskTypes'>): Observable<ClaimOwnershipResult> {
-    const initialCapacity = this.getCapacity();
-    return from(this.getClaimingBatches()).pipe(
-      mergeScan(
-        (accumulatedResult, batch) => {
-          const stopTaskTimer = startTaskTimer();
-          const capacity = Math.min(
-            initialCapacity - accumulatedResult.stats.tasksClaimed,
-            isLimited(batch) ? this.getCapacity(batch.tasksTypes) : this.getCapacity()
-          );
-          // if we have no more capacity, short circuit here
-          if (capacity <= 0) {
-            return of(accumulatedResult);
-          }
-          return from(
-            this.executeClaimAvailableTasks({
-              claimOwnershipUntil,
-              size: capacity,
-              taskTypes: isLimited(batch) ? new Set([batch.tasksTypes]) : batch.tasksTypes,
-            }).then((result) => {
-              const { stats, docs } = accumulateClaimOwnershipResults(accumulatedResult, result);
-              stats.tasksConflicted = correctVersionConflictsForContinuation(
-                stats.tasksClaimed,
-                stats.tasksConflicted,
-                initialCapacity
-              );
-              return { stats, docs, timing: stopTaskTimer() };
-            })
-          );
-        },
-        // initialise the accumulation with no results
-        accumulateClaimOwnershipResults(),
-        // only run one batch at a time
-        1
-      )
-    );
-  }
-
-  private executeClaimAvailableTasks = async ({
-    claimOwnershipUntil,
-    size,
-    taskTypes,
-  }: OwnershipClaimingOpts): Promise<ClaimOwnershipResult> => {
-    const { updated: tasksUpdated, version_conflicts: tasksConflicted } =
-      await this.markAvailableTasksAsClaimed({
-        claimOwnershipUntil,
-        size,
-        taskTypes,
-      });
-
-    const docs = tasksUpdated > 0 ? await this.sweepForClaimedTasks(taskTypes, size) : [];
-
-    this.emitEvents(docs.map((doc) => asTaskClaimEvent(doc.id, asOk(doc))));
-
-    const stats = {
-      tasksUpdated,
-      tasksConflicted,
-      tasksClaimed: docs.length,
-    };
-
-    return {
-      stats,
-      docs,
-    };
-  };
-
-  private isTaskTypeExcluded(taskType: string) {
-    for (const excludedType of this.excludedTaskTypes) {
-      if (minimatch(taskType, excludedType)) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  private async markAvailableTasksAsClaimed({
-    claimOwnershipUntil,
-    size,
-    taskTypes,
-  }: OwnershipClaimingOpts): Promise<UpdateByQueryResult> {
-    const { taskTypesToSkip = [], taskTypesToClaim = [] } = groupBy(
-      this.definitions.getAllTypes(),
-      (type) =>
-        taskTypes.has(type) && !this.isTaskTypeExcluded(type)
-          ? 'taskTypesToClaim'
-          : 'taskTypesToSkip'
-    );
-    const queryForScheduledTasks = mustBeAllOf(
-      // Task must be enabled
-      EnabledTask,
-      // Either a task with idle status and runAt <= now or
-      // status running or claiming with a retryAt <= now.
-      shouldBeOneOf(IdleTaskWithExpiredRunAt, RunningOrClaimingTaskWithExpiredRetryAt)
-    );
-
-    const sort: NonNullable<SearchOpts['sort']> = [SortByRunAtAndRetryAt];
-    const query = matchesClauses(queryForScheduledTasks, filterDownBy(InactiveTasks));
-    const script = updateFieldsAndMarkAsFailed({
-      fieldUpdates: {
-        ownerId: this.taskStore.taskManagerId,
-        retryAt: claimOwnershipUntil,
-      },
-      claimableTaskTypes: taskTypesToClaim,
-      skippedTaskTypes: taskTypesToSkip,
-      unusedTaskTypes: this.unusedTypes,
-      taskMaxAttempts: pick(this.taskMaxAttempts, taskTypesToClaim),
-    });
-
-    const apmTrans = apm.startTransaction(
-      TASK_MANAGER_MARK_AS_CLAIMED,
-      TASK_MANAGER_TRANSACTION_TYPE
-    );
-
-    try {
-      const result = await this.taskStore.updateByQuery(
-        {
-          query,
-          script,
-          sort,
-        },
-        {
-          max_docs: size,
-        }
-      );
-      apmTrans.end('success');
-      return result;
-    } catch (err) {
-      apmTrans.end('failure');
-      throw err;
-    }
-  }
-
-  /**
-   * Fetches tasks from the index, which are owned by the current Kibana instance
-   */
-  private async sweepForClaimedTasks(
-    taskTypes: Set<string>,
-    size: number
-  ): Promise<ConcreteTaskInstance[]> {
-    const claimedTasksQuery = tasksClaimedByOwner(
-      this.taskStore.taskManagerId,
-      tasksOfType([...taskTypes])
-    );
-    const { docs } = await this.taskStore.fetch({
-      query: claimedTasksQuery,
-      size,
-      sort: SortByRunAtAndRetryAt,
-      seq_no_primary_term: true,
-    });
-
-    return docs;
-  }
 }
 
-const emptyClaimOwnershipResult = () => {
-  return {
-    stats: {
-      tasksUpdated: 0,
-      tasksConflicted: 0,
-      tasksClaimed: 0,
-      tasksRejected: 0,
-    },
-    docs: [],
-  };
-};
-
-function accumulateClaimOwnershipResults(
-  prev: ClaimOwnershipResult = emptyClaimOwnershipResult(),
-  next?: ClaimOwnershipResult
-) {
-  if (next) {
-    const { stats, docs, timing } = next;
-    const res = {
-      stats: {
-        tasksUpdated: stats.tasksUpdated + prev.stats.tasksUpdated,
-        tasksConflicted: stats.tasksConflicted + prev.stats.tasksConflicted,
-        tasksClaimed: stats.tasksClaimed + prev.stats.tasksClaimed,
-      },
-      docs,
-      timing,
-    };
-    return res;
-  }
-  return prev;
-}
-
-function isLimited(
+export function isLimited(
   batch: TaskClaimingBatch<BatchConcurrency.Limited | BatchConcurrency.Unlimited, unknown>
 ): batch is LimitedBatch {
   return batch.concurrency === BatchConcurrency.Limited;
 }
+
 function asLimited(tasksType: string): LimitedBatch {
   return {
     concurrency: BatchConcurrency.Limited,
     tasksTypes: tasksType,
   };
 }
+
 function asUnlimited(tasksTypes: Set<string>): UnlimitedBatch {
   return {
     concurrency: BatchConcurrency.Unlimited,
